Use DB in preferences
